arreglado putnbr
cambiado el int por long long para que pueda manejar el INT_MIN y
cambiado el manejo del número para que no de 0 de más

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: paromero <paromero@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/07 11:00:24 by paromero          #+#    #+#             */
-/*   Updated: 2023/11/22 12:00:19 by paromero         ###   ########.fr       */
+/*   Updated: 2023/11/28 11:11:17 by paromero         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,10 +20,12 @@
 # include <stdlib.h>
 # include <stdint.h>
 # include <stdarg.h>
+# include <stdarg.h>
+# include <limits.h>
 
 int		ft_putchar(char c);
 int		ft_putstr(char *str);
-int		ft_putnbr(int n);
+int		ft_putnbr(long long n);
 int		ft_puthex(uintptr_t n, int is_uppercase, int include_prefix);
 int		ft_format(char format, va_list arg);
 int		ft_printf(char	const	*format, ...);
